+++ src/_igraph/vertexseqobject.h
@@ -43,7 +43,7 @@ int igraphmodule_VertexSeq_init(igraphmodule_VertexSeqObject* self,
   PyObject* args, PyObject* kwds);
 void igraphmodule_VertexSeq_dealloc(igraphmodule_VertexSeqObject* self);
 
-int igraphmodule_VertexSeq_sq_length(igraphmodule_VertexSeqObject *self);
+Py_ssize_t igraphmodule_VertexSeq_sq_length(igraphmodule_VertexSeqObject *self);
 
 PyObject* igraphmodule_VertexSeq_find(igraphmodule_VertexSeqObject *self,
   PyObject *args);
